# swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Storage NourriCo - OpenAPI 3.0",
      "description": "This is the Storage API for the micro-services architecture of the NourriCo company.",
      "termsOfService": "",
      "contact": {
        "email": "storage@nourri.co"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.11"
    },
    "externalDocs": {
      "description": "",
      "url": ""
    },
    "servers": [
      {
        "url": "https://localhost:3000"
      }
    ],
    "tags": [
      {
        "name": "product",
        "description": "Manage the products in the storage",
        "externalDocs": {
          "description": "Find out more",
          "url": "http://swagger.io"
        }
      }
    ],
    "paths": {
      "/": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "Show all products",
          "description": "Show all products",
          "operationId": "showProducts",
          "requestBody": {
            "description": "Show all products of storage",
            "required": false
          }
        }
      },
      "/add": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "Add a product",
          "description": "Add a product by array",
          "operationId": "addProduct",
          "requestBody": {
            "description": "Add a product to the storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          }
        }
      },
      "/update": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "update a product",
          "description": "update products by their name in the storage",
          "operationId": "updateProduct",
          "requestBody": {
            "description": "update a product in the storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          }
        }
      },
      "/search": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "search a product",
          "description": "search a product by its name in the storage",
          "operationId": "searchProduct",
          "requestBody": {
            "description": "search a product in the storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          }
        }
      },
      "/updateStock": {
        "get": {
          "tags": [
            "product"
          ],
          "summary": "update stock of a product",
          "description": "update stock of a product by its name in the storage",
          "operationId": "updateStockProduct",
          "requestBody": {
            "description": "update stock of a product in the storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "required": true
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Product": {
          "required": [
            "quantity",
            "name"
          ],
          "type": "object",
          "properties": {
            "quantity": {
              "type": "integer",
              "format": "int64",
              "example": 25
            },
            "name": {
              "type": "string",
              "example": "Computer"
            },
            "type": {
              "type": "string",
              "description": "Product type in storage",
              "example": "furniture",
              "enum": [
                "food",
                "furniture"
              ]
            }
          },
          "xml": {
            "wrapped": true,
            "name": "product"
          }
        },
        "ApiResponse": {
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "type": {
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "xml": {
            "name": "##default"
          }
        }
      },
      "requestBodies": {
        "Product": {
          "description": "Product array that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "UserArray": {
          "description": "List of user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "securitySchemes": {
        "petstore_auth": {
          "type": "oauth2",
          "flows": {
            "implicit": {
              "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
              "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
              }
            }
          }
        },
        "api_key": {
          "type": "apiKey",
          "name": "api_key",
          "in": "header"
        }
      }
    }
  }